Extract location display options into separate component.
